requirements.txt
discord.py-self>=1.9.2
aiohttp>=3.9.0
PyYAML>=6.0
rich>=13.0.0
